dom/security/test/csp/test_data_csp_inheritance.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1381761 - Treating 'data:' documents as unique, opaque origins should still inherit the CSP</title>
  <!-- Including SimpleTest.js so we can use waitForExplicitFinish !-->
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<iframe style="width:100%;" id="testframe"></iframe>

<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

/* Description of the test:
 * We load an iframe using a meta CSP which includes another iframe
 * using a data: URI. We make sure the CSP gets inherited into
 * the data: URI iframe.
 */

window.addEventListener("message", receiveMessage);
function receiveMessage(event) {
  window.removeEventListener("message", receiveMessage);
  // toplevel CSP should apply to data: URI iframe hence resulting
  // in 1 applied policy.
  is(event.data.result, 1,
     "data: URI iframe inherits CSP from including context");
  SimpleTest.finish();
}

document.getElementById("testframe").src = "file_data_csp_inheritance.html";

</script>
</body>
</html>
